@@ -2,30 +2,20 @@
   "name": "jest-demo",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "lint": "eslint src --ext ts,js --fix",
-    "test": "jest --verbose --collectCoverage",
-    "build": "tsc --project tsconfig.build.json"
-  },
-  "dependencies": {
-    "ts-jest": "^29.1.2"
+    "lint": "eslint src/**/*.spec.ts --fix",
+    "test": "jest --verbose --collectCoverage"
   },
   "devDependencies": {
-    "@babel/traverse": "^7.23.7",
-    "@types/jest": "^29.5.11",
-    "@types/node": "^20.11.6",
-    "@typescript-eslint/eslint-plugin": "^6.19.1",
-    "eslint": "^8.56.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-standard-with-typescript": "^43.0.1",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.1",
+    "eslint": "^9.16.0",
+    "globals": "^15.13.0",
     "jest": "^29.7.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0"
   },
-  "commons": [
-    "@babel/traverse 依赖用于为相关依赖提升版本, 低版本有安全问题"
-  ]
+  "packageManager": "pnpm@9.14.4+sha512.c8180b3fbe4e4bca02c94234717896b5529740a6cbadf19fa78254270403ea2f27d4e1d46a08a0f56c89b63dc8ebfd3ee53326da720273794e6200fcf0d184ab"
 }
